requirements.txt
black==23.12.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
idna==3.6
isort==5.13.2
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.1.0
python-dotenv==1.0.0
requests==2.31.0
tenacity==8.2.3
tqdm==4.66.1
urllib3==2.1.0
